+++ src/settings/settings.ts
@@ -620,6 +620,19 @@ export default class InitiativeTrackerSettings extends PluginSettingTab {
         };
         const summary = additionalContainer.createEl("summary");
         new Setting(summary).setHeading().setName("Statuses");
+
+        new Setting(additionalContainer)
+            .setName("Unconscious Status")
+            .setDesc(
+                "Choose a different status to be used as the default Unconscious status."
+            )
+            .addDropdown((d) => {
+                for (const status of this.plugin.data.statuses) {
+                    d.addOption(status.id, status.name);
+                }
+                d.setValue(this.plugin.data.unconsciousId);
+                d.onChange((id) => (this.plugin.data.unconsciousId = id));
+            });
         summary.createDiv("collapser").createDiv("handle");
         const add = new Setting(additionalContainer)
             .setName("Add New Status")
@@ -681,7 +694,34 @@ export default class InitiativeTrackerSettings extends PluginSettingTab {
         const additional = additionalContainer.createDiv("additional");
         for (const status of this.plugin.data.statuses) {
             new Setting(additional)
-                .setName(status.name)
+                .setName(
+                    createFragment((e) => {
+                        const div = e.createDiv("status-name-container");
+                        div.createSpan({ text: status.name });
+
+                        div.createDiv("status-metadata-container");
+                        if (status.resetOnRound) {
+                            setIcon(
+                                div.createDiv({
+                                    attr: {
+                                        "aria-label": "Reset Each Round"
+                                    }
+                                }),
+                                "timer-reset"
+                            );
+                        }
+                        if (status.hasAmount) {
+                            setIcon(
+                                div.createDiv({
+                                    attr: {
+                                        "aria-label": "Has Amount"
+                                    }
+                                }),
+                                "hash"
+                            );
+                        }
+                    })
+                )
                 .setDesc(status.description)
                 .addExtraButton((b) =>
                     b.setIcon("pencil").onClick(() => {
@@ -729,6 +769,9 @@ export default class InitiativeTrackerSettings extends PluginSettingTab {
                             this.plugin.data.statuses.filter(
                                 (s) => s.name != status.name
                             );
+                        if (this.plugin.data.unconsciousId == status.id) {
+                            this.plugin.data.unconsciousId = "Unconscious";
+                        }
                         await this.plugin.saveSettings();
                         this._displayStatuses(additionalContainer);
                     })
@@ -1080,7 +1123,8 @@ class NewPlayerModal extends Modal {
                     this.player.hp = hp;
                     this.player.level = level;
                     this.player.modifier = modifier;
-                    this.player["statblock-link"] = metaData.frontmatter["statblock-link"];
+                    this.player["statblock-link"] =
+                        metaData.frontmatter["statblock-link"];
                     this.display();
                 };
             });
@@ -1266,6 +1310,7 @@ class NewPlayerModal extends Modal {
 
 import { App, ButtonComponent, Modal } from "obsidian";
 import { tracker } from "src/tracker/stores/tracker";
+import { getId } from "src/utils/creature";
 
 export async function confirmWithModal(
     app: App,
@@ -1327,7 +1372,7 @@ addIcon(
 );
 
 class StatusModal extends Modal {
-    status: Condition = { name: null, description: null };
+    status: Condition = { name: null, description: null, id: getId() };
     canceled = false;
     editing: boolean = false;
     original: string;
@@ -1338,14 +1383,15 @@ class StatusModal extends Modal {
             this.original = status.name;
             this.status = {
                 name: status.name,
-                description: status.description
+                description: status.description,
+                id: status.id ?? getId()
             };
         }
     }
     warned = false;
     onOpen() {
         this.titleEl.setText(this.editing ? "Edit Status" : "New Status");
-
+        this.contentEl.empty();
         const name = new Setting(this.contentEl)
             .setName("Name")
             .addText((t) => {
@@ -1384,6 +1430,40 @@ class StatusModal extends Modal {
                 (v) => (this.status.description = v)
             );
         });
+        new Setting(this.contentEl)
+            .setName("Remove Each Round")
+            .setDesc(
+                "This status will be removed from all creatures at the start of a new round."
+            )
+            .addToggle((t) =>
+                t
+                    .setValue(this.status.resetOnRound)
+                    .onChange((v) => (this.status.resetOnRound = v))
+            );
+        new Setting(this.contentEl)
+            .setName("Has Amount")
+            .setDesc(
+                "This status has an amount that can be increased or decreased during combat."
+            )
+            .addToggle((t) =>
+                t.setValue(this.status.hasAmount).onChange((v) => {
+                    this.status.hasAmount = v;
+                    this.onOpen();
+                })
+            );
+        if (this.status.hasAmount) {
+            new Setting(this.contentEl)
+                .setName("Starting Amount")
+                .setDesc("The status will default to this amount when added.")
+                .addText(
+                    (t) =>
+                        (t
+                            .setValue(`${this.status.startingAmount}`)
+                            .onChange((v) => {
+                                this.status.amount = this.status.startingAmount = Number(v);
+                            }).inputEl.type = "number")
+                );
+        }
 
         new ButtonComponent(
             this.contentEl.createDiv("initiative-tracker-cancel")