+++ src/tracker/ui/Updating.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import type { Condition, UpdateLogMessage } from "index";
-    import { setIcon } from "obsidian";
+    import { ExtraButtonComponent, setIcon } from "obsidian";
     import type InitiativeTracker from "src/main";
     import { AC, HP, REMOVE, TAG } from "src/utils";
     import { ConditionSuggestionModal } from "src/utils/suggester";
@@ -8,6 +8,7 @@
 
     import { tracker } from "../stores/tracker";
     const { updating, updateTarget } = tracker;
+    import { Writable, writable } from "svelte/store";
 
     const plugin = getContext<InitiativeTracker>("plugin");
     const hpIcon = (node: HTMLElement) => {
@@ -18,6 +19,10 @@
     };
     const tagIcon = (node: HTMLElement) => {
         setIcon(node, TAG);
+    };
+    let statusBtn: ExtraButtonComponent;
+    const applyStatusIcon = (node: HTMLElement) => {
+        statusBtn = new ExtraButtonComponent(node).setIcon("plus-circle");
     };
     const removeIcon = (node: HTMLElement) => {
         setIcon(node, REMOVE);
@@ -30,15 +35,40 @@
     };
     let damage: string = "";
     let ac: string = "";
-    let status: Condition = null;
-
+    let status: string = null;
+    $: {
+        if (statusBtn) statusBtn.setDisabled(!status);
+    }
+    const statuses = writable<Condition[]>([]);
+    const applyStatus = () => {
+        if (status) {
+            $statuses = [
+                ...$statuses,
+                {
+                    ...plugin.data.statuses.find((s) => s.id == status),
+                },
+            ];
+            status = null;
+        }
+    };
     let modal: ConditionSuggestionModal;
     const suggestConditions = (node: HTMLInputElement) => {
-        modal = new ConditionSuggestionModal(plugin, node);
-        modal.onClose = () => {
-            status = modal.condition;
-            node.focus();
-        };
+        if (!modal) {
+            modal = new ConditionSuggestionModal(
+                plugin.data.statuses
+                    .filter((s) => !$statuses.find((a) => a.id == s.id))
+                    .map((s) => s.id),
+                node
+            );
+            modal.onClose = () => {
+                status = modal.condition;
+                node.focus();
+            };
+        } else {
+            modal.items = plugin.data.statuses
+                .filter((s) => !$statuses.find((a) => a.id == s.id))
+                .map((s) => s.id);
+        }
         modal.open();
     };
     function init(el: HTMLInputElement, target: "hp" | "ac") {
@@ -46,13 +76,15 @@
     }
     const performUpdate = (perform: boolean) => {
         if (perform) {
-            tracker.doUpdate(damage ?? "", status, ac);
+            tracker.doUpdate(damage ?? "", $statuses);
         } else {
             tracker.clearUpdate();
         }
         damage = null;
         status = null;
         ac = null;
+        $statuses = [];
+        modal = null;
         return;
     };
 </script>
@@ -105,39 +137,33 @@
                             saving throw
                         </small>
                     {/if}
-                    <div class="input">
-                        <tag
-                            use:tagIcon
-                            aria-label="Apply status effect to creatures that fail their saving throw"
-                            style="margin: 0 0.2rem 0 0.7rem"
-                        />
-                        <input
-                            type="text"
-                            on:focus={function (evt) {
-                                suggestConditions(this);
-                            }}
-                            on:keydown={function (evt) {
-                                if (["Enter", "Escape"].includes(evt.key)) {
-                                    performUpdate(evt.key == "Enter");
-                                }
-                            }}
+                    <div class="input-status">
+                        <div class="input">
+                            <div
+                                use:tagIcon
+                                aria-label="Apply status effect to creatures that fail their saving throw"
+                                style="margin: 0 0.2rem 0 0.7rem"
+                            />
+                            <input
+                                type="text"
+                                bind:value={status}
+                                on:focus={function (evt) {
+                                    suggestConditions(this);
+                                }}
+                                on:keydown={function (evt) {
+                                    if (["Enter", "Escape"].includes(evt.key)) {
+                                        performUpdate(evt.key == "Enter");
+                                    }
+                                }}
+                            />
+                        </div>
+                        <div
+                            use:applyStatusIcon
+                            aria-label="Add Status"
+                            on:click={applyStatus}
                         />
                     </div>
                 </div>
-            </div>
-            <div class="updating-buttons">
-                <span
-                    use:checkIcon
-                    on:click={() => performUpdate(true)}
-                    style="cursor:pointer"
-                    aria-label="Apply"
-                />
-                <span
-                    use:cancelIcon
-                    on:click={() => performUpdate(false)}
-                    style="cursor:pointer"
-                    aria-label="Cancel"
-                />
             </div>
         {:else}
             <div class="hp-status">
@@ -147,7 +173,7 @@
                 <div class="input">
                     <tag
                         use:acIcon
-                        aria-label="Apply damage, healing(-) or temp HP(t)"
+                        aria-label="Set or (+/-)modify the AC of creatures"
                         style="margin: 0 0.2rem 0 0.7rem"
                     />
                     <input
@@ -168,6 +194,41 @@
             </div>
         {/if}
     </div>
+    {#if $updateTarget == "hp"}
+        <div class="statuses">
+            <table class="updating-creature-table">
+                <thead class="updating-creature-table-header">
+                    <th style="width:100%" class="left">Status</th>
+                    <th
+                        style="cursor:pointer"
+                        class="left"
+                        use:removeIcon
+                        on:click={() => performUpdate(false)}
+                    />
+                </thead>
+                <tbody>
+                    {#each $statuses as status}
+                        <tr class="updating-creature-table-row">
+                            <td>
+                                <span>
+                                    {status.name}
+                                </span>
+                            </td>
+
+                            <td
+                                use:removeIcon
+                                on:click={function (evt) {
+                                    $statuses.remove(status);
+                                    $statuses = $statuses;
+                                }}
+                                style="cursor:pointer"
+                            />
+                        </tr>
+                    {/each}
+                </tbody>
+            </table>
+        </div>
+    {/if}
     {#if plugin.data.beginnerTips}
         <div>
             <small>Multiple creatures can be selected at a time.</small>
@@ -176,29 +237,22 @@
     <div style="margin: 0.5rem">
         <table class="updating-creature-table">
             <thead class="updating-creature-table-header">
-                <th
-                    style="padding:0 0.2rem 0 0; cursor:pointer"
-                    class="left"
-                    use:removeIcon
-                    on:click={() => performUpdate(false)}
-                />
                 <th style="width:100%" class="left">Name</th>
                 {#if $updateTarget == "hp"}
                     <th style="padding:0 0.2rem" class="center">Saved</th>
                     <th style="padding:0 0.2rem" class="center">Resist</th>
                     <th style="padding:0 0.2rem" class="center">Modifier</th>
+                    <th
+                        style="cursor:pointer"
+                        class="left"
+                        use:removeIcon
+                        on:click={() => performUpdate(false)}
+                    />
                 {/if}
             </thead>
             <tbody>
                 {#each [...$updating.entries()] as [creature, update], i}
                     <tr class="updating-creature-table-row">
-                        <td
-                            use:removeIcon
-                            on:click={function (evt) {
-                                tracker.setUpdate(creature, evt);
-                            }}
-                            style="cursor:pointer"
-                        />
                         <td>
                             <span>
                                 {creature.name +
@@ -245,10 +299,31 @@
                                 />
                             </td>
                         {/if}
+                        <td
+                            use:removeIcon
+                            on:click={function (evt) {
+                                tracker.setUpdate(creature, evt);
+                            }}
+                            style="cursor:pointer"
+                        />
                     </tr>
                 {/each}
             </tbody>
         </table>
+    </div>
+    <div class="updating-buttons">
+        <span
+            use:checkIcon
+            on:click={() => performUpdate(true)}
+            style="cursor:pointer"
+            aria-label="Apply"
+        />
+        <span
+            use:cancelIcon
+            on:click={() => performUpdate(false)}
+            style="cursor:pointer"
+            aria-label="Cancel"
+        />
     </div>
 {/if}
 
@@ -257,6 +332,20 @@
         display: flex;
         align-items: center;
         gap: 0.5rem;
+    }
+    :global(.is-disabled) {
+        cursor: not-allowed;
+    }
+    .input-status {
+        display: flex;
+        justify-content: space-between;
+    }
+    .input > div,
+    .input-status > div,
+    td:has(> svg),
+    th:has(> svg) {
+        display: flex;
+        align-items: center;
     }
     .left {
         text-align: left;
@@ -283,5 +372,8 @@
         justify-content: flex-end;
         gap: 1rem;
         margin-right: 1.2rem;
+    }
+    td > input {
+        margin: 0;
     }
 </style>